more tests for sync audio
